moved race test
